Make the LiquidCrystal row offsets uint8_t instead of int
Since these are memory addresses, there is no need to make them signed.
Furthermore, the HD44780 chip supports memory addresses up to 0x67, so
uint8_t shouldbe sufficient.

+++ libraries/LiquidCrystal/src/LiquidCrystal.h
@@ -102,7 +102,7 @@ private:
   uint8_t _initialized;
 
   uint8_t _numlines,_currline;
-  int _row_offsets[4];
+  uint8_t _row_offsets[4];
 };
 
 #endif
